Multi-delete framework ready.
There is an error with the cursor *after* performing the delete, when trying to
quit.  Also, when checkboxes leave the view, the checked status is reset for
some reason.  Finally, a case needs to be added to remove an element from the
arraylist when transitioning from checked to unchecked.  Will the state reset
cause this to happen repeatedly? We'll see!