Conditional private list entries almost working (baby step).
Multidelete set to set all entries to private, because I can't get a non-null
cursor from the database.  Once that's in place, the existing logic should work
and everything should be peachy-keen and hunky-dory.

Multidelete inner class for each element pulled out into its own file as a
non-static class.  This seems like the right thing to do, but there's a chance
it isn't.